+++ client/src/components/games/AvailableGames.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { useContext } from "react";
 import { GamesContext,VenuesContext } from "../../App";
 import { format, differenceInCalendarDays } from 'date-fns';
-import NavBar from "./NavBar";
+import NavBar from "../common/NavBar";
 import { NavLink} from "react-router-dom";
