+++ client/src/components/games/GameDetails.jsx
@@ -2,7 +2,7 @@ import { useContext, useState, useEffect } from "react"
 import { useParams } from "react-router-dom"
 import { GamesContext } from "../../App"
 import { format} from 'date-fns';
-import NavBar from "./NavBar";
+import NavBar from "../common/NavBar";
 
 function GameDetails () {
