+++ client/src/pages/Home.jsx
@@ -1,6 +1,6 @@
-import NavBar from './NavBar';
-import RegisteredVenueList from './registeredVenueList';
-import ActionContainer from './ActionContainer';
+import NavBar from '../components/common/NavBar';
+import RegisteredVenueList from '../components/venues/RegisteredVenueList';
+import ActionContainer from '../components/common/ActionContainer';
 
 
 function Home () {
